pyproject.toml
[tool.poetry]
name = "todo-list"
version = "0.1.0"
description = ""
authors = ["BlaackWizard <kabynks@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
django-ninja = "^1.3.0"
pre-commit = "^4.0.1"
psycopg2-binary = "^2.9.10"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
